DataProcess/zx_python_ui_module_7/requirements.txt
numpy
pillow
pyserial
pydicom
matplotlib
PySide6
opencv-python
